Fix 'double' -> 'float' conversion

+++ src/point_2.h
@@ -28,8 +28,8 @@ class Point_2 {
   Bbox_2 bbox() const { return Bbox_2(_x, _y, _x, _y); }
   /** Distance to other point */
   float dist(const Point_2& p) const {
-    float dx = x() - p.x();
-    float dy = y() - p.y();
+    double dx = x() - p.x();
+    double dy = y() - p.y();
     return sqrt(dx * dx + dy * dy);
   }
